backend/requirements.txt
# LegalEdge AI Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
alembic==1.12.1

# OpenAI & AI
openai==0.28.1

# PDF Processing
PyPDF2==3.0.1

# Data Validation
pydantic==2.5.0

# CORS & Security
python-multipart==0.0.6

# Analytics (Optional)
posthog==3.1.0

# Payment Processing (Optional)
stripe==7.6.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

